backend/pyproject.toml
[tool.poetry]
name = "git-recap"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
pygithub = "^2.1.1"
pydantic = "^2.5.1"
pydantic-settings = "^2.1.0"
pyyaml = "^6.0.1"
openai = "^1.3.5"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
wemake-python-styleguide = "^0.18.0"
mypy = "^1.7.0"
pytest-cov = "^4.1.0"
coverage-conditional-plugin = "^0.9.0"
freezegun = "^1.2.2"
pydocstyle = "^6.3.0"
types-pyyaml = "^6.0.12.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
